Move the MaterialInputOnWhite style to its component
