+++ src/components/scenes/DefaultFiatSettingScene.js
@@ -9,13 +9,12 @@ import { setDefaultFiatRequest } from '../../actions/SettingsActions'
 import s from '../../locales/strings.js'
 import { getDefaultFiat } from '../../modules/Settings/selectors'
 import Text from '../../modules/UI/components/FormattedText/FormattedText.ui.js'
-import { MaterialInputOnWhite } from '../../styles/components/FormFieldStyles.js'
 import { THEME } from '../../theme/variables/airbitz'
 import { type Dispatch, type State as ReduxState } from '../../types/reduxTypes.js'
 import type { FlatListItem, GuiFiatType } from '../../types/types.js'
 import { scale } from '../../util/scaling.js'
 import { getSupportedFiats } from '../../util/utils'
-import { FormField } from '../common/FormField.js'
+import { FormField, MaterialInputOnWhite } from '../common/FormField.js'
 import { SceneWrapper } from '../common/SceneWrapper.js'
 
 type StateProps = {
